Split out test_pep_number
